Reconcillation between main and dev + Pushing docs changes (#11)
* Add CI filtering, .ciignore, CONTRIBUTING, and Third-Party Notices (#2)

* Added contributing guidelines, another workflow to prevent unecessary build invocation, misc stuff mostly

* Added credits to third party libs

* Give name to filter workflow

* Should fix CI filtering issues now, hopefully

* Logging files which trigger CI build for debugging purposes + .gitignore is added to .ciignore (#4)

* Optimized epoll event loop (#5)

* Logging files which trigger CI build for debugging purposes + .gitignore is added to .ciignore (Forgot to fetch dev so this unecessary commit again, nvm for now)

* Optimized epoll event loop, epoll_ctl is not called in hot path now, few other minor changes

* Forms implemented, refactored parts of code (#6)

* Form support added for in memory forms, need to fix support for SSR

* Basic forms are now complete. Can validate, sanitize, render fields. Fixed few bugs in template engine and refactored http response

* Basic docs + code refactoring (#8)

* Basic documentation is ready, enough to get user onboard, more things will be added in future

* Automated CI to publish latest docs to github pages

* Updated docs + Code refactoring (#10)

* Completed core concepts section in docs, moving onto api reference next

* Updated docs, added API References for request, response, routing and middleware. Slight code refactoring in places

+++ include/http/helper.hpp
@@ -38,8 +38,9 @@ HttpCallbackType MakeHttpCallbackFromLambda(Lambda&& cb)
 
 // vvv Middleware Stuff vvv
 template<typename T>
-inline MiddlewareEntry MakeMiddlewareEntry(T&& funcOrEntry, std::uint8_t handle = MiddlewareType::LINEAR)
-{
+inline MiddlewareEntry MakeMiddlewareEntry(
+    T&& funcOrEntry, std::uint8_t handle = static_cast<std::uint8_t>(MiddlewareType::LINEAR)
+) {
     using Request = WFX::Http::HttpRequest;
     using RawT    = std::decay_t<T>;
